refactor(tools): extract registerStyleguideTools into tools/styleguide.js
Phase C item 5: moves 8 styleguide tools (setup_prose_styleguide_config,
get_prose_styleguide_config, summarize_prose_styleguide_config,
bootstrap_prose_styleguide_config, update_prose_styleguide_config,
preview_prose_styleguide_config_update, check_prose_styleguide_drift,
setup_prose_styleguide_skill) out of index.js. isPathCandidateInsideSyncDir
is passed through toolContext. Removes prose-styleguide, prose-styleguide-drift,
and prose-styleguide-skill imports from index.js; retains STYLEGUIDE_CONFIG_BASENAME
which is still used by get_runtime_config.
